+++ specs/css/selectorFinder.spec.js
@@ -15,68 +15,68 @@ describe('CSS-Selector Finder Testing', function () {
     });
 
     it('should find "row-wrapper"', function () {
-      var foundCssSelectors = s.findCssSelectors('.row-wrapper {');
+      var foundCssSelectors = s.find('.row-wrapper {');
       expect(foundCssSelectors._class).toEqual(['row-wrapper']);
     });
 
     it('should find "row-wrapper and mytest"', function () {
-      var foundCssSelectors = s.findCssSelectors('.row-wrapper.mytest {');
+      var foundCssSelectors = s.find('.row-wrapper.mytest {');
       expect(foundCssSelectors._class).toEqual(['row-wrapper', 'mytest']);
     });
 
     it('should find "hallo"', function () {
-      var foundCssSelectors = s.findCssSelectors('td.hallo { 2. Prevent i,');
+      var foundCssSelectors = s.find('td.hallo { 2. Prevent i,');
       expect(foundCssSelectors._class).toEqual(['hallo']);
     });
 
     it('should find "[]"', function () {
-      var foundCssSelectors = s.findCssSelectors('2. Prevent iOS and IE,');
+      var foundCssSelectors = s.find('2. Prevent iOS and IE,');
       expect(foundCssSelectors._class).toEqual([]);
     });
 
     it('should find "[]"', function () {
-      var foundCssSelectors = s.findCssSelectors('src: url("../ok/0c-da2.et?#ie");');
+      var foundCssSelectors = s.find('src: url("../ok/0c-da2.et?#ie");');
       expect(foundCssSelectors._class).toEqual([]);
     });
 
     it('should find "myClass"', function () {
-      var foundCssSelectors = s.findCssSelectors('.myClass:before {');
+      var foundCssSelectors = s.find('.myClass:before {');
       expect(foundCssSelectors._class).toEqual(['myClass']);
     });
 
     it('should find "page--xyz" and "testClass1"', function () {
-      var foundCssSelectors = s.findCssSelectors('body:not(.page--xyz) .testClass1 {');
+      var foundCssSelectors = s.find('body:not(.page--xyz) .testClass1 {');
       expect(foundCssSelectors._class).toEqual(['page--xyz', 'testClass1']);
     });
 
     it('should find "[]"', function () {
-      var foundCssSelectors = s.findCssSelectors('url("//fast.fonts.net/");');
+      var foundCssSelectors = s.find('url("//fast.fonts.net/");');
       expect(foundCssSelectors._class).toEqual([]);
     });
 
     it('should find "[]"', function () {
-      var foundCssSelectors = s.findCssSelectors('padding: 1em .71429em .92857em 0;');
+      var foundCssSelectors = s.find('padding: 1em .71429em .92857em 0;');
       expect(foundCssSelectors._class).toEqual([]);
     });
 
     it('should find "[]"', function () {
       var t = '/*! normalize-scss | MIT/GPLv2 License | bit.ly/xyz.css */';
-      var foundCssSelectors = s.findCssSelectors(t);
+      var foundCssSelectors = s.find(t);
       expect(foundCssSelectors._class).toEqual([]);
     });
 
     it('should find "xyz_abc" and "icon" ', function () {
-      var foundCssSelectors = s.findCssSelectors('.xyz_abc .icon,');
+      var foundCssSelectors = s.find('.xyz_abc .icon,');
       expect(foundCssSelectors._class).toEqual(['xyz_abc', 'icon']);
     });
 
     it('should find "xyz_abc" only one time ', function () {
-      var foundCssSelectors = s.findCssSelectors('.xyz_abc .icon.xyz_abc,');
+      var foundCssSelectors = s.find('.xyz_abc .icon.xyz_abc,');
       expect(foundCssSelectors._class).toEqual(['xyz_abc', 'icon']);
     });
 
     it('should find "language" and "deactive" ', function () {
-      var foundCssSelectors = s.findCssSelectors('.language span.deactive{');
+      var foundCssSelectors = s.find('.language span.deactive{');
       expect(foundCssSelectors._class).toEqual(['language', 'deactive']);
     });
 
@@ -92,35 +92,35 @@ describe('CSS-Selector Finder Testing', function () {
     });
 
     it('should find "myId" ', function () {
-      var foundCssSelectors = s.findCssSelectors('#myId');
+      var foundCssSelectors = s.find('#myId');
       expect(foundCssSelectors._id).toEqual(['myId']);
     });
 
     it('should find "onlyId" ', function () {
-      var foundCssSelectors = s.findCssSelectors('#onlyId.icon');
+      var foundCssSelectors = s.find('#onlyId.icon');
       expect(foundCssSelectors._id).toEqual(['onlyId']);
     });
 
     it('should find "myId" "testID" ', function () {
-      var foundCssSelectors = s.findCssSelectors('#myId.icon #testID');
+      var foundCssSelectors = s.find('#myId.icon #testID');
       expect(foundCssSelectors._id).toEqual(['myId', 'testID']);
     });
 
     it('should find "[]"', function () {
       var t = '   border: 1px solid #c0c0c0;';
-      var foundCssSelectors = s.findCssSelectors(t);
+      var foundCssSelectors = s.find(t);
       expect(foundCssSelectors._id).toEqual([]);
     });
 
     it('should find "[]"', function () {
       var t = '   border: 1px solid #c0c0c0 #ff0;';
-      var foundCssSelectors = s.findCssSelectors(t);
+      var foundCssSelectors = s.find(t);
       expect(foundCssSelectors._id).toEqual([]);
     });
 
     it('should find "[]"', function () {
       var t = '   background: -webkit-gradient(linear, left top,  from(#ebf2f8), to(#d8e6ef));';
-      var foundCssSelectors = s.findCssSelectors(t);
+      var foundCssSelectors = s.find(t);
       expect(foundCssSelectors._id).toEqual([]);
     });
 
@@ -135,23 +135,23 @@ describe('CSS-Selector Finder Testing', function () {
 
     it('should find "[]"', () => {
       var comment = '/*';
-      var foundCssSelectors = s.findCssSelectors(comment);
+      var foundCssSelectors = s.find(comment);
       comment = '* .xyz abc.html and test.html';
-      foundCssSelectors = s.findCssSelectors(comment);
+      foundCssSelectors = s.find(comment);
       comment = '*/';
-      foundCssSelectors = s.findCssSelectors(comment);
+      foundCssSelectors = s.find(comment);
       expect(foundCssSelectors._class).toEqual([]);
     });
 
     it('should find "[]"', () => {
       var comment = '/* .xyz abc.html and test.html';
-      var foundCssSelectors = s.findCssSelectors(comment);
+      var foundCssSelectors = s.find(comment);
       expect(foundCssSelectors._class).toEqual([]);
     });
 
     it('should find "[]"', () => {
       var comment = '/* .xyz abc.html and test.html */';
-      var foundCssSelectors = s.findCssSelectors(comment);
+      var foundCssSelectors = s.find(comment);
       expect(foundCssSelectors._class).toEqual([]);
     });
   });